rest api functions operational (need to fix a db function)

+++ Server/src/main/java/org/server/MealSessionInfo.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class MealSessionInfo {
     public KitchenConstraint kitchenConstraints;
-    public List<Integer> recipeIDs;
+    public List<Long> recipeIDs;
     //Should be emails
     public List<String> includedFriends;
